requirements.txt
Flask==3.0.0
Flask-SocketIO==5.3.6
python-dotenv==1.0.0
redis==5.0.3
pycryptodome==3.20.0
eventlet==0.36.1
gunicorn==22.0.0
